Add support for events beginning with silence
Some events beginning with silence would trigger
the Save() call in the Recorder. To better support
these events, the recorder will now give a 1second
window of grace period for a non-silent audio sample
to be received, but only for the first non-silent sample.
Afterwards, the grace window is reduced back to 250ms.